Update to 1.7.2 and Buildcraft 6.0.x
With the introduction of bc 6 or "Next-Gen" the included objects and
pipes will likely change.
---The folder structure was changed and adapted.
---The kinesis system was (attempted at being) updated for the new MJ...
systems and powerhandler.
---Deleted some pipes that Buildcraft made native (Stripes Pipe, Bounce
Pipe, Pump Pipe, etc)
---Updated for new forge item systems
---Completely re-wrote the Gate Settings Duplicator (see below for
usage)
--Added windmills
--Added Ender Storage
----Added Ender Extraction Pipe
----Made Insertion Pipes insert into ender chests

Changed the following recipes to adapt for the removal of the Autarchic
Gate:
Valve Pipe- replaced Autarchic Gate with Basic Gate
Directed Kinesis Pipe- replace Autarchic Gate with Basic Gate
Gate Settings Duplicator- replace Autarchic Gate with pulse chipset

This is beta. As far as I'm aware everything works as flawlessly as
possible with the provided tools, and the testing is rather thorough but
only by three people from three computers. All of the computers were
running Windows 7 Ultimate 64-bit with Java 7 update 50, and have at
least quad core CPUs and high-mid-range GPUs with 8-16GB of RAM in LAN
on a managed gigabit switch. If your specs match, good for you. If not,
sorry, we don't have what you do, so it's not tested yet.

Any and all problems are encouraged to be reported, but they might
already be known, and I might already be working on them.

With that, enjoy your pipes.

Division Pipe:
Splits stacks into a specified size 1-8, incremented with a wrench. If
the stack has a remainder, the moving item will be reversed.

This has been tested with a 64-stack of arrows and a stripes pipe. I
recommend trying it.

Gate Settings Duplicator:
Will replicate ONLY the settings of a gate. The gate must be (by design)
of the same material, logic type, and have the same expansion(s).
Usage- Right-click the PIPE with the configured gate attached, then
right-click another PIPE with a gate of the same type (defined above).
The settings are not lost or overriden regardless of usage. Right-click
a Redstone Block to wipe the settings

Windmill:
Recipe: a plus with iron ingots, and diamond gears in each four corners
Based on placement it can produce very little or a large amount of
energy.
Placement rules:
Must be placed on 2 fence posts (or nether fence)
must not have any opaque blocks within 2 blocks at its height or 1 below
Can NOT push energy to the east (over the blades, they don't rotate)

Windmill energy rules are grouped into two sections: Plains and Hills.

Windmills gain two types of boosts: Biome and Height.

Plains Rules:
If a windmill is placed in a biome with low height variation (flat), it
will follow the Plains rule set.
Bonuses:
Biome:
Windmills gain a bonus for simply being placed in the plains section,
the amount of bonus depends on the biome. Flatter is better.

Height:
Windmills placed as close as possible to sea level will provide the
maximum height boost. Sea level assumed 64 -> + 2 for fence posts -> the
windmill block should be at 66.

Hills Rules:
Biome:
Windmills gain no bonus from biome placement in hills, but can make up
for it and still achieve maximum output with Height.

Height:
The higher a windmill is placed in hills, the greater the bonus, capping
at a height of 124 meters.

All windmills gain an extra a boost in the rain.

The maximum output for windmills is 1.375 MJ/t without rain and 1.5 MJ/t
with rain.

Ender Storage:
Simply an Ender Chest for Buildcraft. Adding an insertion pipe to an
ender chest allows putting items in. Right-clicking on the insertion
pipe when connected opens the bc Ender Chest for easy of use and
debugging purposes, for that "where'd all my items go?" moment.

Insertion: The Insertion Pipe will choose to add items to the ender
chest before anything else. If you don't like that, don't put an ender
chest next to an insertion pipe. It is the only pipe that will connect
to them now.

Ender Extraction Pipe:
Recipe: Extraction Pipe and an ender pearl.
Usage:
Stole/borrowed the gui from the emerald pipe in buildcraft, along with
all the logic and everything but the item extraction, which only pulls
from the ender storage. I doubt BC will sue me, and they can have it. I
just wanted a way to transport items 6000 blocks without making all that
pipe and then keeping it loaded.
Yes, this works across dimensions. If you're smart, you'll find a way to
bucket lava from the nether to use in the overworld. That was your hint.

Power it with an engine, it'll extract just like an emerald but doesn't
need to be connected to a chest.

Tested:
-All Item Pipes
-All Fluid Pipes
-All Kinesis Pipes
-Gate Settings Duplicator
-Windmills!!
-Ender Storage (all aspects)

Everything above has been tested in both singleplayer and multiplayer.